Move `Partition` package to util
