+++ grakn-graql/src/main/java/ai/grakn/graql/internal/gremlin/spanningtree/ChuLiuEdmonds.java
@@ -18,7 +18,7 @@
 
 package ai.grakn.graql.internal.gremlin.spanningtree;
 
-import ai.grakn.graql.internal.gremlin.spanningtree.datastructure.Partition;
+import ai.grakn.graql.internal.util.Partition;
 import ai.grakn.graql.internal.gremlin.spanningtree.graph.DirectedEdge;
 import ai.grakn.graql.internal.gremlin.spanningtree.graph.WeightedGraph;
 import ai.grakn.graql.internal.gremlin.spanningtree.util.Pair;
