+++ grakn-graql/src/main/java/ai/grakn/graql/internal/gremlin/spanningtree/EdgeQueueMap.java
@@ -19,7 +19,7 @@
 package ai.grakn.graql.internal.gremlin.spanningtree;
 
 import ai.grakn.graql.internal.gremlin.spanningtree.datastructure.FibonacciQueue;
-import ai.grakn.graql.internal.gremlin.spanningtree.datastructure.Partition;
+import ai.grakn.graql.internal.util.Partition;
 import ai.grakn.graql.internal.gremlin.spanningtree.graph.DirectedEdge;
 import ai.grakn.graql.internal.gremlin.spanningtree.util.Pair;
 import ai.grakn.graql.internal.gremlin.spanningtree.util.Weighted;
