+++ grakn-graql/src/main/java/ai/grakn/graql/internal/util/Partition.java
@@ -14,9 +14,10 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
+ *
  */
 
-package ai.grakn.graql.internal.gremlin.spanningtree.datastructure;
+package ai.grakn.graql.internal.util;
 
 import com.google.common.collect.Maps;
